requirements.txt
behave
selenium
pytest
allure-behave  # optional for reports